requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
autopep8==1.5.4
certifi==2021.10.8
cffi==1.15.0
coverage==6.3.2
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.12
django-rest-passwordreset==1.2.1
django-tastypie==0.14.4
djangorestframework==3.13.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
lxml==4.8.0
outcome==1.1.0
psycopg2==2.9.3
pycodestyle==2.6.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2021.3
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
wsproto==1.1.0
